requirements.txt
Werkzeug>=2.2.0,<3.0.0
Flask==2.2.5
flask-cors==3.0.10
boto3==1.28.3
pymongo==4.4.1
python-dotenv==1.0.0
tensorflow
Pillow>=9.0.0  # Updated for security patches
numpy>=1.21.0  # More compatible with recent TensorFlow versions